# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental API - GES-CARS-2026",
        "description": "API RESTful pour la gestion de location de voitures",
        "contact": {
            "email": "admin@ges-cars.ma"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1"
        }
    ],
    "paths": {
...
        "/vignettes": {
            "get": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Liste des vignettes",
                "operationId": "5d0e87814d31a010f7636d8b3bb5f1e9",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_paid",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Créer une vignette",
                "operationId": "83016eae0216ba9120f6ac8b144d399f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "year",
                                    "issue_date",
                                    "expiry_date",
                                    "amount"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "issue_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "expiry_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    },
                                    "payment_reference": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vignette created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vignettes/{id}": {
            "get": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Détails d'une vignette",
                "operationId": "23ae16ad815dde366d8802fec7d3af0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Modifier une vignette",
                "operationId": "f12cfdc3b0ca4b721fbeac7e940f1af3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Supprimer une vignette",
                "operationId": "31931e30668385b8402ceca7e19a7249",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vignettes/expired": {
            "get": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Vignettes expirées",
                "operationId": "6f51214b0ad77a00c0bd47775f9731b2",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vignettes/unpaid": {
            "get": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Vignettes non payées",
                "operationId": "2ba25f522a94c03925ca3cef1d73ff19",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vignettes/{id}/mark-paid": {
            "patch": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Marquer vignette comme payée",
                "operationId": "d44c908a8601b657ce36c1c9f9579380",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "bank_transfer",
                                            "online"
                                        ]
                                    },
                                    "payment_reference": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marked as paid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vignettes/{id}/document": {
            "post": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Upload document vignette",
                "operationId": "57c1f881712e9e874031c174b9ba2412",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "document": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uploaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vignettes/{id}/payment-proof": {
            "post": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Upload preuve de paiement",
                "operationId": "7ab57f4c71cbef320de87ebb99679112",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "payment_proof": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uploaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vignettes/{id}/media/{mediaId}": {
            "delete": {
                "tags": [
                    "Vignettes"
                ],
                "summary": "Supprimer un média",
                "operationId": "865561fd21b3dbcb50290991d7012151",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mediaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrer le token JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}